refactor: Zod 스키마를 사용하여 jti를 포함한 JWT 타입 정의 및 적용 #20
