+++ src/modules/auth/decorators/auth-user.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { AccessTokenPayload } from '@/shared/jwt/jwt.service.interface';
+import { AccessTokenPayload } from '@/shared/jwt/jwt.payload.schema';
 import { Request } from 'express';
 
 // createParamDecorator를 사용해 @AuthUser() 데코레이터를 만듭니다.
